address pr feedback
* move tests to specific network module
