+++ bindings/rust/integration/src/lib.rs
@@ -1,6 +1,9 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+#[cfg(all(feature = "network-tests", test))]
+mod network;
+
 #[cfg(test)]
 mod tests {
     use s2n_tls::{
